Added a local SQLite database
The database is completely standalone. It does not require anything to
be installed or configured. It runs completely from a dll file living in
the project folder.

I am currently just putting random junk into the database to prove it
works. The busness logic of the app will need to be added.